requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-rest-passwordreset==1.3.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==15.3.4
idna==3.4
joblib==1.2.0
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
toposort==1.7
tzdata==2022.7
urllib3==1.26.13
